fix: consolidate AddCustomIdentity() into BuildingBlocks.API
