--- ConsumerApi/AspNetCoreIdentityCustomizations/CustomLookupNormalizer.cs
@@ -1,16 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-
-namespace ConsumerApi.AspNetCoreIdentityCustomizations;
-
-public class CustomLookupNormalizer : ILookupNormalizer
-{
-    public string? NormalizeName(string? name)
-    {
-        return name?.Trim().Normalize().ToUpperInvariant();
-    }
-
-    public string? NormalizeEmail(string? email)
-    {
-        return email?.Trim().Normalize().ToUpperInvariant();
-    }
-}
